Don't overwrite the ID field if it is found to be a duplicate
Manually setting the sequential ID column is risky business
however you slice it. Leaving things as they were opens a hole
for an ID to be changed automatically, which is also very bad.
If the sequential ID column is never updated manually, there
certainly will not be duplicates. If it is updated manually, we
make no guarantees about it's uniqueness.