Update gemspec to exclude test assets from compiled gem
We noticed that the .gem file unnecessarily includes the test directory, including a 3 megabyte test.log file after installation. I've updated the gemspec line with the latest convention from `bundle gem`'s generator, and removed the `test_files` directive which as far as I can tell never served a purpose, as is no longer included in the gemspec reference.